+++ apps/starter-e2e/playwright.config.ts
@@ -1,0 +1,26 @@
+import { workspaceRoot } from '@nx/devkit';
+import { nxE2EPreset } from '@nx/playwright/preset';
+import { defineConfig, devices } from '@playwright/test';
+
+// For CI you can override BASE_URL to point at a deployed environment.
+const baseURL = process.env['BASE_URL'] ?? 'http://localhost:4221';
+
+export default defineConfig({
+  ...nxE2EPreset(__filename, { testDir: './src' }),
+  use: {
+    baseURL,
+    trace: 'on-first-retry',
+  },
+  webServer: {
+    command: 'pnpm exec nx run starter:serve',
+    url: 'http://localhost:4221',
+    reuseExistingServer: true,
+    cwd: workspaceRoot,
+  },
+  projects: [
+    {
+      name: 'chromium',
+      use: { ...devices['Desktop Chrome'] },
+    },
+  ],
+});